replace not found page with a warning banner
